+++ lib/prop_odds_api/api/v1.ex
@@ -1,7 +1,27 @@
 defmodule PropOddsAPI.Api.V1 do
+  @moduledoc """
+  This module provides functionality to interact with V1 of the PropOdds API.
+  """
   alias PropOddsAPI.Connection
   import PropOddsAPI.RequestBuilder
 
+  @doc """
+  Fetches fantasy lines for a given game and market.
+
+  ## Parameters
+
+  - `connection`: The `Tesla.Env.client()` used to make the HTTP request.
+  - `game_id`: The unique identifier for the game, as a `String.t()`.
+  - `market`: The market type for which the fantasy lines are being requested, as a `String.t()`.
+  - `api_key`: The API key used for authorization, as a `String.t()`.
+  - `opts`: Optional parameters provided as a keyword list. Can include `:active_only` and `:end_datetime` to filter the results.
+
+  ## Returns
+
+  - `{:ok, PropOddsAPI.Model.FantasyBooks.t()}`: On success, returns the fantasy books data.
+  - `{:ok, PropOddsAPI.Model.HttpValidationError.t()}`: If there's a validation error with the HTTP request.
+  - `{:error, Tesla.Env.t()}`: On failure, returns the error within the Tesla environment.
+  """
   @spec get_fantasy_lines(Tesla.Env.client(), String.t(), String.t(), String.t(), keyword()) ::
           {:ok, PropOddsAPI.Model.FantasyBooks.t()}
           | {:ok, PropOddsAPI.Model.HttpValidationError.t()}
@@ -28,6 +48,23 @@ defmodule PropOddsAPI.Api.V1 do
     ])
   end
 
+  @doc """
+  Retrieves a snapshot of fantasy data for a given league and market.
+
+  ## Parameters
+
+  - `connection`: The `Tesla.Env.client()` used to make the HTTP request.
+  - `league`: The identifier for the league, as a `String.t()`.
+  - `market`: The market type for which the fantasy snapshot is being requested, as a `String.t()`.
+  - `api_key`: The API key used for authorization, as a `String.t()`.
+  - `opts`: Optional parameters provided as a keyword list. Can include `:active_only` to filter the results.
+
+  ## Returns
+
+  - `{:ok, PropOddsAPI.Model.FantasySnapshot.t()}`: On success, returns the fantasy snapshot data.
+  - `{:ok, PropOddsAPI.Model.HttpValidationError.t()}`: If there's a validation error with the HTTP request.
+  - `{:error, Tesla.Env.t()}`: On failure, returns the error within the Tesla environment.
+  """
   @spec get_fantasy_snapshot(Tesla.Env.client(), String.t(), String.t(), String.t(), keyword()) ::
           {:ok, PropOddsAPI.Model.FantasySnapshot.t()}
           | {:ok, PropOddsAPI.Model.HttpValidationError.t()}
@@ -53,6 +90,22 @@ defmodule PropOddsAPI.Api.V1 do
     ])
   end
 
+  @doc """
+  Retrieves a league-wide fantasy snapshot.
+
+  ## Parameters
+
+  - `connection`: The `Tesla.Env.client()` used to make the HTTP request.
+  - `league`: The identifier for the league, as a `String.t()`.
+  - `api_key`: The API key used for authorization, as a `String.t()`.
+  - `opts`: Optional parameters provided as a keyword list. Can include `:active_only` to filter the results.
+
+  ## Returns
+
+  - `{:ok, PropOddsAPI.Model.FantasyLeagueSnapshot.t()}`: On success, returns the fantasy league snapshot data.
+  - `{:ok, PropOddsAPI.Model.HttpValidationError.t()}`: If there's a validation error with the HTTP request.
+  - `{:error, Tesla.Env.t()}`: On failure, returns the error within the Tesla environment.
+  """
   @spec get_league_fantasy_snapshot(Tesla.Env.client(), String.t(), String.t(), keyword()) ::
           {:ok, PropOddsAPI.Model.FantasyLeagueSnapshot.t()}
           | {:ok, PropOddsAPI.Model.HttpValidationError.t()}
@@ -78,6 +131,23 @@ defmodule PropOddsAPI.Api.V1 do
     ])
   end
 
+  @doc """
+  Fetches odds for a given game and market.
+
+  ## Parameters
+
+  - `connection`: The `Tesla.Env.client()` used to make the HTTP request.
+  - `game_id`: The unique identifier for the game, as a `String.t()`.
+  - `market`: The market type for which the odds are being requested, as a `String.t()`.
+  - `api_key`: The API key used for authorization, as a `String.t()`.
+  - `opts`: Optional parameters provided as a keyword list. Can include `:active_only` and `:end_datetime` to filter the results.
+
+  ## Returns
+
+  - `{:ok, PropOddsAPI.Model.SportsBooks.t()}`: On success, returns the odds from various sportsbooks.
+  - `{:ok, PropOddsAPI.Model.HttpValidationError.t()}`: If there's a validation error with the HTTP request.
+  - `{:error, Tesla.Env.t()}`: On failure, returns the error within the Tesla environment.
+  """
   @spec get_odds(Tesla.Env.client(), String.t(), String.t(), String.t(), keyword()) ::
           {:ok, PropOddsAPI.Model.SportsBooks.t()}
           | {:ok, PropOddsAPI.Model.HttpValidationError.t()}