+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+<a name="v0.0.2"></a>
+## v0.0.2
+
+> 2024-01-10
+
+### Chores
+
+* document api functions properly
+* add tests for usage, games, markets
+* fix config
+
 <a name="v0.0.1"></a>
 ## v0.0.1